Add files via upload
The goal of this project was to build a robust machine learning pipeline to classify energy production levels into categories — High, Medium, and Low — using historical data from the years 2019 and 2021. The dataset provides key insights into energy trends and serves as the basis for supervised learning to assist in decision-making for energy planning and sustainability efforts.